requirements.txt
numpy>=1.24.3
Pillow>=12.1.0
scikit_learn>=1.3.2
torch>=2.4.1
torchvision>=0.19.1
tqdm>=4.67.1
transformers>=4.46.3
